pingvin: Re-work normal directions in cross-section (z-y plane)
GMSH creates 2D grids in the x-y plane, and I require those grids
in the z-y plane for pingvin. Based on how I was using the GMSH output
in the x-y plane and converting to z-y plane, I had ended up with
unit normals pointing in the negative x direction. That was ok, so
long as I was careful: fluxes needed a negative sense to them, etc.

In this commit, I reworked the interpretation of quads when building
faces in the z-y plane. Essentially, this means reversing the order
of quad points when reading from the GMSH 2D grid. This gives us
quad unit normals that are (1, 0, 0), ie. point in the positive
x-direction.

This change had several knock on effects: the assembly of
hexahedrons from quads needed reworking, and the streamwise
fluxes needed correction (now defined in the positive sense).