pingvin: Change to using conserved mass-energy-entropy flux average
We observed problems with extracting properties via one-d averaging
for quantities that should have remained unchanged in an isentropic
part of the compression fields. It turns out the one-d averaging
based on conservation of mass-momentum-energy can introduce entropy
due to the "mixing" of momenta streams. To remedy this, switched
to using the conserved mass-energy-entropy flux average.

See Baurle and Gaffney paper for more details.